move the home tests to their folders
